requirements.txt
twitchio
python-dotenv
requests
pillow
customtkinter
flask